test(chain): split indexed authority snapshots
