+++ packages/chain/test/context-graph-authority-snapshot.unit.test.ts
@@ -36,10 +36,6 @@ interface AuthorityEvidence {
   readonly staticCalls: Array<readonly [bigint, { blockTag: number }]>;
   readonly deploymentReads: Array<readonly [string, string, string]>;
   readonly readOptions: Array<Readonly<{ policy?: string; signal?: AbortSignal }>>;
-  readonly indexRanges: Array<readonly [number, number]>;
-  readonly indexTopicSets: string[][];
-  readonly indexAddresses: string[];
-  readonly indexInvalidations: number[];
 }
 
 interface EvmAuthorityHarness {
@@ -49,10 +45,8 @@ interface EvmAuthorityHarness {
   advanceAuthorityHead(): void;
   replaceCachedAnchor(): void;
   replaceFinalizedHead(): void;
-  replaceAuthorityFork(): void;
   holdCurrentStateRead(): Readonly<{ entered: Promise<void>; release(): void }>;
   holdBlockRead(tag: string | number): Readonly<{ entered: Promise<void>; release(): void }>;
-  holdIndexPageRead(): Readonly<{ entered: Promise<void>; release(): void }>;
   setPublishAuthorityAccountId(value: unknown): void;
   rotateContextGraphStorage(): void;
 }
@@ -61,39 +55,14 @@ function makeEvmAuthorityAdapter(
   options: {
     reorg?: boolean;
     providerRangeLimit?: number;
-    sharedIndex?: boolean;
-    deactivated?: boolean;
   } = {},
 ): EvmAuthorityHarness {
-  let authorityIndexRecord: Readonly<{ token: number; value: unknown | null }> | undefined;
-  const indexInvalidations: number[] = [];
-  const authorityIndexStore = {
-    load: async () => authorityIndexRecord,
-    compareAndSwap: async (
-      _scope: string,
-      expectedToken: number | undefined,
-      value: unknown,
-    ) => {
-      if (authorityIndexRecord?.token !== expectedToken) return undefined;
-      const nextToken = expectedToken === undefined ? 1 : expectedToken + 1;
-      authorityIndexRecord = { token: nextToken, value };
-      return nextToken;
-    },
-    invalidate: async (_scope: string, expectedToken: number) => {
-      if (authorityIndexRecord?.token !== expectedToken) return undefined;
-      const nextToken = expectedToken + 1;
-      authorityIndexRecord = { token: nextToken, value: null };
-      indexInvalidations.push(nextToken);
-      return nextToken;
-    },
-  };
   const adapter: any = new EVMChainAdapter({
     rpcUrl: 'http://127.0.0.1:1',
     hubAddress: GOVERNANCE,
     privateKey: `0x${'11'.repeat(32)}`,
     allowNoAdminSigner: true,
     chainId: 'evm:31337',
-    ...(options.sharedIndex ? { localContextGraphAuthorityIndexStore: authorityIndexStore } : {}),
   });
   adapter.initialized = true;
   adapter.init = async () => {};
@@ -104,23 +73,14 @@ function makeEvmAuthorityAdapter(
     staticCalls: [] as Array<readonly [bigint, { blockTag: number }]>,
     deploymentReads: [] as Array<readonly [string, string, string]>,
     readOptions: [],
-    indexRanges: [],
-    indexTopicSets: [],
-    indexAddresses: [],
-    indexInvalidations,
   };
 
   let finalizedNumber = 30;
   let finalizedHash = FINALIZED_HASH;
   let cachedAnchorReplaced = false;
-  let replacementAuthorityFork = false;
   let currentReadGate: PromiseWithResolvers<void> | undefined;
   let blockReadGate: Readonly<{
     tag: string | number;
-    entered: PromiseWithResolvers<void>;
-    release: PromiseWithResolvers<void>;
-  }> | undefined;
-  let indexPageReadGate: Readonly<{
     entered: PromiseWithResolvers<void>;
     release: PromiseWithResolvers<void>;
   }> | undefined;
@@ -136,9 +96,6 @@ function makeEvmAuthorityAdapter(
       AUTHORITY,
       7n,
     ])],
-    ContextGraphDeactivated: options.deactivated
-      ? [event(24, 0, `0x${'bc'.repeat(32)}`, [9n])]
-      : [],
     Transfer: [
       event(10, 0, CREATION_HASH, [ethers.ZeroAddress, SECOND_MEMBER, 9n]),
       event(15, 0, `0x${'99'.repeat(32)}`, [SECOND_MEMBER, OWNER, 9n]),
@@ -151,11 +108,11 @@ function makeEvmAuthorityAdapter(
     AgentParticipantRemoved: [event(23, 0, `0x${'bb'.repeat(32)}`, [9n, SECOND_MEMBER])],
   };
   const current = Object.assign(
-    [OWNER, [MEMBER, OWNER], 0n, !options.deactivated, 0n, 1n, 0n, AUTHORITY, 7n],
+    [OWNER, [MEMBER, OWNER], 0n, true, 0n, 1n, 0n, AUTHORITY, 7n],
     {
       owner: OWNER,
       participantAgents: [MEMBER, OWNER],
-      active: !options.deactivated,
+      active: true,
       accessPolicy: 1n,
       publishPolicy: 0n,
       publishAuthority: AUTHORITY,
@@ -236,152 +193,6 @@ function makeEvmAuthorityAdapter(
       };
     },
     getNetwork: async () => ({ chainId: 31337n }),
-    getLogs: async (filter: {
-      address: string;
-      fromBlock: number;
-      toBlock: number;
-      topics: string[][];
-    }) => {
-      expect(filter.address).toBe(GOVERNANCE);
-      evidence.indexAddresses.push(filter.address);
-      evidence.indexRanges.push([filter.fromBlock, filter.toBlock]);
-      evidence.indexTopicSets.push(filter.topics[0] ?? []);
-      const gate = indexPageReadGate;
-      if (gate !== undefined) {
-        indexPageReadGate = undefined;
-        gate.entered.resolve();
-        await gate.release.promise;
-      }
-      const namedArgs = (values: readonly unknown[], names: Record<string, unknown>) => (
-        Object.assign([...values], names)
-      );
-      const forkOwner = replacementAuthorityFork ? MEMBER : SECOND_MEMBER;
-      const forkParticipants = replacementAuthorityFork
-        ? [MEMBER]
-        : [MEMBER, SECOND_MEMBER];
-      return [
-        {
-          blockNumber: 10,
-          blockHash: CREATION_HASH,
-          index: 1,
-          parsed: {
-            name: 'ContextGraphCreated',
-            args: namedArgs([
-              9n,
-              forkOwner,
-              NAME_HASH,
-              forkParticipants,
-              0n,
-              1n,
-              0n,
-              AUTHORITY,
-              7n,
-            ], {
-              contextGraphId: 9n,
-              owner: forkOwner,
-              nameHash: NAME_HASH,
-              participantAgents: forkParticipants,
-              accessPolicy: 1n,
-              publishPolicy: 0n,
-              publishAuthority: AUTHORITY,
-              publishAuthorityAccountId: 7n,
-            }),
-          },
-        },
-        {
-          blockNumber: 10,
-          blockHash: CREATION_HASH,
-          index: 0,
-          parsed: {
-            name: 'Transfer',
-            args: namedArgs([ethers.ZeroAddress, forkOwner, 9n], {
-              from: ethers.ZeroAddress, to: forkOwner, tokenId: 9n,
-            }),
-          },
-        },
-        {
-          blockNumber: 15,
-          blockHash: `0x${'99'.repeat(32)}`,
-          index: 0,
-          parsed: {
-            name: 'Transfer',
-            args: namedArgs([SECOND_MEMBER, OWNER, 9n], {
-              from: SECOND_MEMBER, to: OWNER, tokenId: 9n,
-            }),
-          },
-        },
-        {
-          blockNumber: 20,
-          blockHash: POLICY_HASH,
-          index: 0,
-          parsed: {
-            name: 'PublishPolicyUpdated',
-            args: namedArgs([9n, 0n, SECOND_AUTHORITY, 9n], {
-              contextGraphId: 9n,
-              publishPolicy: 0n,
-              publishAuthority: SECOND_AUTHORITY,
-              publishAuthorityAccountId: 9n,
-            }),
-          },
-        },
-        {
-          blockNumber: 21,
-          blockHash: POLICY_HASH,
-          index: 0,
-          parsed: {
-            name: 'PublishAuthorityUpdated',
-            args: namedArgs([9n, AUTHORITY, 7n], {
-              contextGraphId: 9n,
-              newAuthority: AUTHORITY,
-              newAuthorityAccountId: 7n,
-            }),
-          },
-        },
-        ...[
-          ['AgentParticipantAdded', 22, `0x${'aa'.repeat(32)}`, OWNER],
-          ['AgentParticipantRemoved', 23, `0x${'bb'.repeat(32)}`, SECOND_MEMBER],
-        ].map(([name, blockNumber, hash, agent]) => ({
-          blockNumber,
-          blockHash: hash,
-          index: 0,
-          parsed: {
-            name,
-            args: namedArgs([9n, agent], { contextGraphId: 9n, agent }),
-          },
-        })),
-        ...(options.deactivated ? [{
-          blockNumber: 24,
-          blockHash: `0x${'bc'.repeat(32)}`,
-          index: 0,
-          parsed: {
-            name: 'ContextGraphDeactivated',
-            args: namedArgs([9n], { contextGraphId: 9n }),
-          },
-        }] : []),
-        {
-          blockNumber: 33,
-          blockHash: NEXT_POLICY_HASH,
-          index: 0,
-          parsed: {
-            name: 'PublishPolicyUpdated',
-            args: namedArgs([9n, 1n, ethers.ZeroAddress, 0n], {
-              contextGraphId: 9n,
-              publishPolicy: 1n,
-              publishAuthority: ethers.ZeroAddress,
-              publishAuthorityAccountId: 0n,
-            }),
-          },
-        },
-      ].filter((entry) => {
-        if (
-          replacementAuthorityFork
-          && entry.parsed.name !== 'ContextGraphCreated'
-          && !(entry.parsed.name === 'Transfer' && entry.blockNumber === 10)
-        ) return false;
-        return Number(entry.blockNumber) >= filter.fromBlock
-          && Number(entry.blockNumber) <= filter.toBlock;
-      });
-    },
   };
   adapter.contracts = {
     contextGraphStorage: { connect: () => contract },
@@ -426,11 +237,6 @@ function makeEvmAuthorityAdapter(
       finalizedHash = REPLACEMENT_FINALIZED_HASH;
       cachedAnchorReplaced = true;
     },
-    replaceAuthorityFork: () => {
-      finalizedHash = REPLACEMENT_FINALIZED_HASH;
-      cachedAnchorReplaced = true;
-      replacementAuthorityFork = true;
-    },
     holdCurrentStateRead: () => {
       const entered = Promise.withResolvers<void>();
       const release = Promise.withResolvers<void>();
@@ -447,12 +253,6 @@ function makeEvmAuthorityAdapter(
       blockReadGate = { tag, entered, release };
       return { entered: entered.promise, release: release.resolve };
     },
-    holdIndexPageRead: () => {
-      const entered = Promise.withResolvers<void>();
-      const release = Promise.withResolvers<void>();
-      indexPageReadGate = { entered, release };
-      return { entered: entered.promise, release: release.resolve };
-    },
     setPublishAuthorityAccountId: (value) => {
       current.publishAuthorityAccountId = value;
       current[8] = value;
@@ -462,220 +262,6 @@ function makeEvmAuthorityAdapter(
 }
 
 describe('RFC-64 Context Graph authority snapshots', () => {
-  it('uses one combined contract-wide log request per page when the durable index is wired', async () => {
-    const { adapter, evidence } = makeEvmAuthorityAdapter({ sharedIndex: true });
-
-    const snapshot = await adapter.getContextGraphAuthoritySnapshot(9n);
-    expect(snapshot).toMatchObject({
-      contextGraphId: '9',
-      owner: OWNER,
-      active: true,
-      accessPolicy: 1,
-      publishPolicy: 0,
-      publishAuthority: AUTHORITY,
-      publishAuthorityAccountId: '7',
-      participantAgents: [OWNER, MEMBER],
-      ownershipEra: '1',
-      policyVersion: '3',
-      rosterVersion: '3',
-      sourceBlockNumber: '21',
-    });
-    expect(evidence.indexRanges).toEqual([[7, 16], [17, 26], [27, 30]]);
-    expect(evidence.indexAddresses).toEqual(Array(3).fill(GOVERNANCE));
-    expect(evidence.filters).toEqual([]);
-    expect(evidence.indexTopicSets).toEqual(Array(3).fill([
-      'topic:ContextGraphCreated',
-      'topic:ContextGraphDeactivated',
-      'topic:Transfer',
-      'topic:PublishPolicyUpdated',
-      'topic:PublishAuthorityUpdated',
-      'topic:AgentParticipantAdded',
-      'topic:AgentParticipantRemoved',
-    ]));
-    expect(evidence.staticCalls).toEqual([]);
-
-    await adapter.getContextGraphAuthoritySnapshot(9n);
-    expect(evidence.indexRanges).toHaveLength(3);
-    expect(evidence.staticCalls).toEqual([]);
-  });
-
-  it('rejects an indexed reorg fence then invalidates and rebuilds the replacement fork', async () => {
-    const harness = makeEvmAuthorityAdapter({ sharedIndex: true });
-    const stabilization = harness.holdBlockRead(30);
-    const stale = harness.adapter.getContextGraphAuthoritySnapshot(9n);
-
-    await stabilization.entered;
-    harness.replaceAuthorityFork();
-    stabilization.release();
-    await expect(stale).rejects.toThrow('anchor changed');
-
-    await expect(harness.adapter.getContextGraphAuthoritySnapshot(9n)).resolves.toMatchObject({
-      contextGraphId: '9',
-      owner: MEMBER,
-      participantAgents: [MEMBER],
-      ownershipEra: '0',
-      policyVersion: '0',
-      rosterVersion: '0',
-    });
-    expect(harness.evidence.indexInvalidations).toEqual([4]);
-    expect(harness.evidence.indexRanges).toEqual([
-      [7, 16], [17, 26], [27, 30],
-      [7, 16], [17, 26], [27, 30],
-    ]);
-  });
-
-  it.each([
-    ['finalized head', (harness: EvmAuthorityHarness) => harness.holdBlockRead('finalized')],
-    ['stabilization fence', (harness: EvmAuthorityHarness) => harness.holdBlockRead(30)],
-  ] as const)('keeps caller cancellation bound during the indexed %s read', async (
-    _stage,
-    hold,
-  ) => {
-    const harness = makeEvmAuthorityAdapter({ sharedIndex: true });
-    const adapter = harness.adapter as any;
-    adapter.readTipProvider = async (
-      _label: string,
-      read: (provider: EvmAuthorityHarness['provider']) => Promise<unknown>,
-      options: Readonly<{ signal?: AbortSignal }>,
-    ) => {
-      harness.evidence.readOptions.push(options);
-      const pending = read(harness.provider);
-      if (options.signal === undefined) return pending;
-      return new Promise((resolve, reject) => {
-        const onAbort = () => reject(options.signal!.reason);
-        options.signal!.addEventListener('abort', onAbort, { once: true });
-        void pending.then(resolve, reject).finally(() => {
-          options.signal!.removeEventListener('abort', onAbort);
-        });
-      });
-    };
-    const gate = hold(harness);
-    const abort = new AbortController();
-    const pending = harness.adapter.getContextGraphAuthoritySnapshot(9n, {
-      signal: abort.signal,
-    });
-    await gate.entered;
-    abort.abort(new Error('snapshot caller left'));
-    await expect(pending).rejects.toThrow('snapshot caller left');
-    expect(harness.evidence.readOptions[0]?.signal).toBe(abort.signal);
-    gate.release();
-  });
-
-  it('detaches one cancelled waiter without aborting its shared indexed page read', async () => {
-    const harness = makeEvmAuthorityAdapter({ sharedIndex: true });
-    const adapter = harness.adapter as any;
-    adapter.readTipProvider = async (
-      _label: string,
-      read: (provider: EvmAuthorityHarness['provider']) => Promise<unknown>,
-      options: Readonly<{ signal?: AbortSignal }>,
-    ) => {
-      const pending = read(harness.provider);
-      if (options.signal === undefined) return pending;
-      return new Promise((resolve, reject) => {
-        const onAbort = () => reject(options.signal!.reason);
-        options.signal!.addEventListener('abort', onAbort, { once: true });
-        void pending.then(resolve, reject).finally(() => {
-          options.signal!.removeEventListener('abort', onAbort);
-        });
-      });
-    };
-    const gate = harness.holdIndexPageRead();
-    const abort = new AbortController();
-    const cancelled = harness.adapter.getContextGraphAuthoritySnapshot(9n, {
-      signal: abort.signal,
-    });
-    await gate.entered;
-    const survivor = harness.adapter.getContextGraphAuthoritySnapshot(9n);
-    abort.abort(new Error('one waiter left'));
-    await expect(cancelled).rejects.toThrow('one waiter left');
-    gate.release();
-    await expect(survivor).resolves.toMatchObject({ contextGraphId: '9' });
-  });
-
-  it('fails over when a provider cannot revalidate the durable authority anchor', async () => {
-    const { adapter, provider, advanceAuthorityHead } = makeEvmAuthorityAdapter({
-      sharedIndex: true,
-    });
-    await adapter.getContextGraphAuthoritySnapshot(9n);
-    advanceAuthorityHead();
-
-    const attempts: string[] = [];
-    const lagging = {
-      ...provider,
-      getBlock: async (tag: string | number) => {
-        if (tag === 30) return null;
-        return (provider.getBlock as (block: string | number) => Promise<unknown>)(tag);
-      },
-    };
-    (adapter as any).readTipProvider = async (
-      _label: string,
-      read: (selected: typeof lagging) => Promise<unknown>,
-      options: Readonly<{ isRetryable?: (error: unknown) => boolean }>,
-    ) => {
-      try {
-        attempts.push('lagging');
-        return await read(lagging);
-      } catch (error) {
-        expect(options.isRetryable?.(error)).toBe(true);
-        attempts.push('healthy');
-        return read(provider as typeof lagging);
-      }
-    };
-
-    await expect(adapter.getContextGraphAuthoritySnapshot(9n)).resolves.toMatchObject({
-      policyVersion: '4',
-      sourceBlockNumber: '33',
-    });
-    expect(attempts).toEqual(['lagging', 'healthy']);
-  });
-
-  it('fails over when a provider cannot supply an intermediate page anchor', async () => {
-    const { adapter, provider } = makeEvmAuthorityAdapter({ sharedIndex: true });
-    const attempts: string[] = [];
-    const nonArchive = {
-      ...provider,
-      getBlock: async (tag: string | number) => {
-        if (tag === 16) return null;
-        return (provider.getBlock as (block: string | number) => Promise<unknown>)(tag);
-      },
-    };
-    (adapter as any).readTipProvider = async (
-      _label: string,
-      read: (selected: typeof nonArchive) => Promise<unknown>,
-      options: Readonly<{ isRetryable?: (error: unknown) => boolean }>,
-    ) => {
-      try {
-        attempts.push('non-archive');
-        return await read(nonArchive);
-      } catch (error) {
-        expect(options.isRetryable?.(error)).toBe(true);
-        attempts.push('healthy');
-        return read(provider as typeof nonArchive);
-      }
-    };
-
-    await expect(adapter.getContextGraphAuthoritySnapshot(9n)).resolves.toMatchObject({
-      contextGraphId: '9',
-      policyVersion: '3',
-    });
-    expect(attempts).toEqual(['non-archive', 'healthy']);
-  });
-
-  it('materializes deactivation from the shared event index without a point read', async () => {
-    const { adapter, evidence } = makeEvmAuthorityAdapter({
-      sharedIndex: true,
-      deactivated: true,
-    });
-
-    await expect(adapter.getContextGraphAuthoritySnapshot(9n)).resolves.toMatchObject({
-      contextGraphId: '9',
-      active: false,
-      owner: OWNER,
-      participantAgents: [OWNER, MEMBER],
-    });
-    expect(evidence.staticCalls).toEqual([]);
-  });
-
   it('reads one stable finalized EVM generation and derives monotonic epochs', async () => {
     const { adapter, evidence, advanceAuthorityHead } = makeEvmAuthorityAdapter();
     const snapshot = await adapter.getContextGraphAuthoritySnapshot(9n);
